another warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@51699 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/zip/pkzipvfs.cpp
@@ -566,8 +566,8 @@ static bool rtZipPkzipReaderScanEocd(const uint8_t *pu8Buf, size_t cb, int *piPo
 {
     if (cb < 4)
         return false;
-    int i;
-    for (i = cb - 4; i >= 0; --i)
+    ssize_t i;
+    for (i = (ssize_t)cb - 4; i >= 0; --i)
         if (*(uint32_t*)(pu8Buf + i) == RTZIPPKZIPENDOFCENTRDIRREC_MAGIC)
         {
             *piPos = i;
